backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
google-api-python-client>=2.150.0
google-generativeai>=0.8.0
python-dotenv>=1.0.1
httpx>=0.27.0
pydantic>=2.10.0
